Splitup in exchange_, file_ and telegram_handling.py files
